test(pptx): keep decoder regressions free of self-dependencies
